requirements.txt
mcp[cli]>=0.1.0
rdkit>=2023.3.1
pillow>=10.0.0
